requirements.txt
# Core GW libraries
gwpy>=3.0.0
pycbc>=2.0.0
bilby>=2.0.0

# ML/AI stack
torch>=2.0.0
nflows>=0.14
scikit-learn>=1.3.0
xgboost>=1.7.0
numpy>=1.24.0
scipy>=1.10.0

# Experiment management
hydra-core>=1.3.0
wandb>=0.15.0

# Utilities
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
tqdm>=4.65.0
pytest>=7.4.0
h5py>=3.8.0
astropy>=5.0.0
requests>=2.28.0
pyyaml>=6.0
